Improved connectTo()/connectFrom() on Transforms
connectTo() and connectFrom() are much more liberal when used with asymmetric transforms: they now require only the connecting data types to match (i.e. the producer's output matches the consumer's input). This allows for a Transform<int, float> to be connected to a Transform<float, float>. Prior to this, Transform<int, float> was not symmetric and thus not allowed to be connected to <float, float>.