Add new transform: Tee class
Tee is a numeric passthrough transform that always passes the value unchanged. It also optionally applies a linear transform of the form y = mx + b, and then passes the y value to an external function.  This external function can be used, for example, to output the  y value in the form of a frequency-modulated signal on a GPIO pin. The Tee class is so called because it inserts a tee, or tap, into the chain of value Producers/Consumers.