Prevent duplicate PUT requests
Fixes the strange timing problem when a PUT request is issued by a remote switch, but the the physical switch controller sends its current status before it answers the PUT request. This mod will also prevent future issues with the server being overwhelmed with requests as this only lets any one PUT request instance to make a single request at a time.