Make Debounce a header-only template
I need Debounce<SystemStatus> for the wifi disconnection watchdog
and it wasn't possibly to instantiate with this setup.

I made a quick test with PlatformIO's inspection tool. I had a random
test program which used DebounceBool and DebounceInt of some
input. The RAM use was 51.0 KB and Flash use was 1015.4 KB. After
changing the transform to header-only templates, RAM use remained
the same but flash use dropped to 1011.0 KB.