TaskQueueProducer constructor requires a ReactESP pointer
There was a major bug in the previous TaskQueueProducer implementation:
values were always received in the main task context. This commit adds
a new constructor argument in which a subtask app pointer can be
provided, allowing queued items to be consumed in a subtask.

A constructor variant without the argument is also provided for
backwards compatibility.