Update curveinterpolator.cpp
Prevent overun of curve producing erroneous values.

For example in a simple curve with two points (100,0) and (500,1)  representing inputs between 100 and 500  producing outputs 0 to 1 ... but any values outside the range eg 99 or 501 produce  arbitary results, for the majority of applications (eg a temperature sender or a fuel level sender) limiting the output to the min or max curve valies makes more sense.