Add new observers to the end of the observer list
This make connect_to and O(N) operation with respect to the
number of existing observers, but the call order of the observers
is now the same as the insertion order. Also, adding new observers
usually happens in the setup stage of program execution, in which
the added performance penalty is not relevant.